Move fixure to conftest file
